backend/requirements.txt
# Essential Django dependencies
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
psycopg2-binary==2.9.9
redis==5.0.1
gunicorn==21.2.0
python-decouple==3.8
djoser==2.2.2
djangorestframework-simplejwt==5.3.0
whitenoise==6.6.0
Pillow==10.1.0

# Simplified AI dependencies (will be enhanced later)
openai==1.51.0
numpy==1.26.2
